requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.3.1
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
ImageIO==2.37.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.1
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.10.2
numpy==2.1.3
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==5.29.5
pyarrow==21.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.4
rich==14.1.0
rpds-py==0.28.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-gradcam==0.1.0
termcolor==3.1.0
tf-keras-vis==0.8.7
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
